[deckhouse] delete stale module at goroutine
